Fixed SA1402 rule
