+++ server/Controllers/BanDto.cs
@@ -1,0 +1,9 @@
+namespace Server.Controllers
+{
+    public class BanDto
+    {
+        public string Ban { get; set; }
+
+        public string PlayerName { get; set; }
+    }
+}
