+++ server/Controllers/ExplorationCardsController.cs
@@ -28,18 +28,4 @@ namespace Server.Controllers
             return objectivesFromDb.Select(fromDb => new ExplorationCardDto(fromDb)).OrderBy(f => f.PlanetType);
         }
     }
-
-    public class ExplorationCardDto : Exploration
-    {
-        public ExplorationCardDto(Exploration exploration)
-        {
-            this.Slug = exploration.Slug;
-            this.GameVersion = exploration.GameVersion;
-            this.PlanetType = exploration.PlanetType;
-            this.NumberOfCards = exploration.NumberOfCards;
-            this.Resources = exploration.Resources;
-            this.Influence = exploration.Influence;
-            this.TechSkip = exploration.TechSkip;
-        }
-    }
 }
