+++ server/Controllers/ObjectivesController.cs
@@ -31,17 +31,4 @@ namespace Server.Controllers
             return objectivesFromDb.Select(fromDb => new ObjectiveDto(fromDb));
         }
     }
-
-    public class ObjectiveDto : Objective
-    {
-        public ObjectiveDto(Objective objective)
-        {
-            this.Slug = objective.Slug;
-            this.GameVersion = objective.GameVersion;
-            this.Points = objective.Points;
-            this.Secret = objective.Secret;
-            this.When = objective.When;
-            this.Reward = objective.Reward;
-        }
-    }
 }
