+++ server/Controllers/RelicsController.cs
@@ -28,13 +28,4 @@ namespace Server.Controllers
             return objectivesFromDb.Select(fromDb => new RelicDto(fromDb));
         }
     }
-
-    public class RelicDto : Card
-    {
-        public RelicDto(Relic relic)
-        {
-            this.Slug = relic.Slug;
-            this.GameVersion = relic.GameVersion;
-        }
-    }
 }
