+++ server/Controllers/ScorableObjectiveDto.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+namespace Server.Controllers
+{
+    public class ScorableObjectiveDto
+    {
+        public string Slug { get; set; }
+
+        public List<string> ScoredBy { get; set; }
+    }
+}
