+++ server/Controllers/SessionDto.cs
@@ -221,138 +221,4 @@ namespace Server.Controllers
             }).ToList();
         }
     }
-
-    public class ScorableObjectiveDto
-    {
-        public string Slug { get; set; }
-
-        public List<string> ScoredBy { get; set; }
-    }
-
-    public class PlayerDto
-    {
-        public PlayerDto()
-        {
-            this.AtTable = -1;
-        }
-
-        public string PlayerName { get; set; }
-
-        public string Faction { get; set; }
-
-        public string Color { get; set; }
-
-        public bool Speaker { get; set; }
-
-        public int AtTable { get; set; }
-
-        public static IEnumerable<PlayerDto> GetPlayers(SessionDto session)
-        {
-            var factionPicks = session.Draft?.Picks?.Where(p => p.Type == "faction") ?? new PickedPayload[0];
-            var tablePicks = session.Draft?.Picks?.Where(p => p.Type == "tablePosition") ?? new PickedPayload[0];
-
-            var picks = session.Factions.Select(faction =>
-            {
-                var decapitalizedFaction = faction;
-                decapitalizedFaction = char.ToLower(decapitalizedFaction[0]) + decapitalizedFaction.Substring(1);
-                var playerName = factionPicks.FirstOrDefault(fp => fp.Pick == faction)?.PlayerName;
-                var tablePick = tablePicks.FirstOrDefault(tp => tp.PlayerName == playerName)?.Pick;
-
-                return new PlayerDto
-                {
-                    Faction = faction,
-                    PlayerName = playerName,
-                    Color = session.Colors?.GetValueOrDefault(faction) ?? session.Colors?.GetValueOrDefault(decapitalizedFaction),
-                    Speaker = playerName != null && session.Draft?.Speaker == playerName,
-                    AtTable = int.Parse(tablePick ?? "-1"),
-                };
-            });
-
-            var speaker = session.Draft?.Speaker;
-
-            if (!string.IsNullOrEmpty(speaker) && tablePicks.Any())
-            {
-                var ordered = tablePicks.OrderBy(tp => int.Parse(tp.Pick));
-                var duplicated = ordered.Concat(ordered).ToList();
-                var speakerIndex = duplicated.FindIndex(a => a.PlayerName == speaker);
-
-                var inOrderAfterSpeaker = duplicated.Skip(speakerIndex).Take(picks.Count());
-
-                return inOrderAfterSpeaker.Select(orderedPick => picks.First(pick => pick.PlayerName == orderedPick.PlayerName));
-            }
-
-            return picks;
-        }
-    }
-
-    public class BanDto
-    {
-        public string Ban { get; set; }
-
-        public string PlayerName { get; set; }
-    }
-
-    public class DraftDto
-    {
-        public DraftDto()
-        {
-        }
-
-        public DraftDto(Session session)
-        {
-            var orderedEvents = session.Events.OrderBy(e => e.HappenedAt);
-            var gameStartEvent = orderedEvents.FirstOrDefault(e => e.EventType == nameof(GameStarted));
-            var gameStartOptions = GameStarted.GetPayload(gameStartEvent).Options;
-
-            var banEvents = session.Events.Where(e => e.EventType == nameof(Banned));
-            var bans = banEvents.SelectMany(b =>
-            {
-                var payload = Banned.GetPayload(b);
-                return payload.Bans.Select(f => new BanDto { Ban = f, PlayerName = payload.PlayerName });
-            }).ToArray();
-            var pickEvents = orderedEvents.Where(e => e.EventType == nameof(Picked));
-            var banOrder = JsonConvert.DeserializeObject<int[]>(orderedEvents.FirstOrDefault(e => e.EventType == "PlayerOrder")?.SerializedPayload ?? "[]");
-            var orderEvent = orderedEvents.LastOrDefault(e => e.EventType == "PlayerOrder");
-
-            this.Order = JsonConvert.DeserializeObject<int[]>(orderEvent?.SerializedPayload ?? "[]");
-            this.Phase = ((gameStartOptions?.Bans ?? false) && banEvents.Count() < banOrder.Count()) ? "bans" :
-              (pickEvents.Count() < this.Order.Count() ? "picks" : "speaker");
-            this.InitialPool = gameStartOptions?.InitialPool;
-            this.Players = gameStartOptions?.Players ?? new string[0];
-            this.BansPerRound = gameStartOptions?.BansPerRound ?? 1;
-            this.Bans = bans;
-            this.Picks = pickEvents.Select(Picked.GetPayload).ToArray();
-            this.ActivePlayerIndex = this.Phase == "bans" ? banEvents.Count() : pickEvents.Count();
-
-            var speakerSelectedEvent = orderedEvents.LastOrDefault(e => e.EventType == nameof(SpeakerSelected));
-            if (speakerSelectedEvent != null)
-            {
-                this.Speaker = SpeakerSelected.GetPayload(speakerSelectedEvent).SpeakerName;
-            }
-
-            var speakerPickedEvent = pickEvents.LastOrDefault(e => Picked.GetPayload(e.SerializedPayload).Type == "speaker");
-            if (speakerPickedEvent != null)
-            {
-                this.Speaker = Picked.GetPayload(speakerPickedEvent.SerializedPayload).PlayerName;
-            }
-        }
-
-        public string[] InitialPool { get; set; }
-
-        public string[] Players { get; set; }
-
-        public BanDto[] Bans { get; set; }
-
-        public PickedPayload[] Picks { get; set; }
-
-        public int BansPerRound { get; set; }
-
-        public string Phase { get; set; }
-
-        public int[] Order { get; set; }
-
-        public int ActivePlayerIndex { get; set; }
-
-        public string Speaker { get; set; }
-    }
 }
