+++ server/Domain/GameEvents/AgendaVotedOn.cs
@@ -46,15 +46,4 @@ namespace Server.Domain
             await this.repository.SaveChangesAsync();
         }
     }
-
-    public class AgendaVotedOnPayload
-    {
-        public string Slug { get; set; }
-
-        public VoteResult Result { get; set; }
-
-        public string Election { get; set; }
-
-        public AgendaType Type { get; set; }
-    }
 }
