+++ server/Domain/GameEvents/CommitDraftPayload.cs
@@ -1,0 +1,7 @@
+namespace Server.Domain
+{
+    public class CommitDraftPayload
+    {
+        public string[] Factions { get; set; }
+    }
+}
