+++ server/Domain/GameEvents/Extensions.cs
@@ -1,0 +1,29 @@
+using System.Collections.Generic;
+using System.Security.Cryptography;
+
+namespace Server.Domain
+{
+    internal static class Extensions
+    {
+        // https://stackoverflow.com/questions/273313/randomize-a-listt
+        public static void Shuffle<T>(this IList<T> list)
+        {
+            RNGCryptoServiceProvider provider = new RNGCryptoServiceProvider();
+            int n = list.Count;
+            while (n > 1)
+            {
+                byte[] box = new byte[1];
+                do
+                {
+                    provider.GetBytes(box);
+                }
+                while (!(box[0] < n * (byte.MaxValue / n)));
+                int k = box[0] % n;
+                n--;
+                T value = list[k];
+                list[k] = list[n];
+                list[n] = value;
+            }
+        }
+    }
+}
