+++ server/Domain/GameEvents/FactionsShuffledPayload.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Server.Domain
+{
+    internal class FactionsShuffledPayload
+    {
+        public List<string> Factions { get; set; }
+    }
+}
