+++ server/Domain/GameEvents/LawRemovedPayload.cs
@@ -1,0 +1,7 @@
+namespace Server.Domain
+{
+    public class LawRemovedPayload
+    {
+        public string Slug { get; set; }
+    }
+}
