+++ server/Domain/GameEvents/ObjectiveAddedPayload.cs
@@ -1,0 +1,7 @@
+namespace Server.Domain
+{
+    public class ObjectiveAddedPayload
+    {
+        public string Slug { get; set; }
+    }
+}
