+++ server/Domain/GameEvents/ObjectiveDeletedPayload.cs
@@ -1,0 +1,7 @@
+namespace Server.Domain
+{
+    internal class ObjectiveDeletedPayload
+    {
+        public string Slug { get; set; }
+    }
+}
