+++ server/Domain/GameEvents/ObjectiveDescoredPayload.cs
@@ -1,0 +1,11 @@
+namespace Server.Domain
+{
+    public class ObjectiveDescoredPayload
+    {
+        public string Slug { get; set; }
+
+        public string Faction { get; set; }
+
+        public int Points { get; set; }
+    }
+}
