+++ server/Domain/GameEvents/ObjectiveScoredPayload.cs
@@ -1,0 +1,11 @@
+namespace Server.Domain
+{
+    public class ObjectiveScoredPayload
+    {
+        public string Slug { get; set; }
+
+        public string Faction { get; set; }
+
+        public int Points { get; set; }
+    }
+}
