+++ server/Domain/GameEvents/Picked.cs
@@ -146,15 +146,4 @@ namespace Server.Domain
             return session.Events.Count(e => e.EventType == nameof(Picked));
         }
     }
-
-    public class PickedPayload
-    {
-        public string Pick { get; set; }
-
-        public string Type { get; set; } // "faction", "tablePosition"
-
-        public int PlayerIndex { get; set; }
-
-        public string PlayerName { get; set; }
-    }
 }
