+++ server/Domain/GameEvents/SpeakerSelectedPayload.cs
@@ -1,0 +1,9 @@
+namespace Server.Domain
+{
+    public class SpeakerSelectedPayload
+    {
+        public int SpeakerIndex { get; set; }
+
+        public string SpeakerName { get; set; }
+    }
+}
