+++ server/Domain/GameEvents/VictoryPointsUpdatedPayload.cs
@@ -1,0 +1,13 @@
+namespace Server.Domain
+{
+    public class VictoryPointsUpdatedPayload
+    {
+        public string Faction { get; set; }
+
+        public int Points { get; set; }
+
+        public VictoryPointSource Source { get; set; }
+
+        public string Context { get; set; }
+    }
+}
